feat: Add debug to understand the code
It doesn't work great, as implemented it doesn't generate optimal results
and there are rounding errors. The obvious non-optimal result is when
num-points is 1, the optimal result is placing it in the center instead
it's always at (0,0) because that is candidates.first(). The algorithm
never adjusts previously thus it cannot be optimal for many other
sizes. Such as num-points 3 where is returns a right triangle and it should
probably be an equal lateral triangle.

Anyway, interesting experinece.